Add files via upload
You can still run automorphism() as an interface. Alternatively, you can directly run faster() (which is the automorphism finder), which I prefer if I am running a lot of similar posets over and over. There are some commented examples at the bottom of the file. faster() takes O(n!) time in the worst case because posets with n disjoint nodes have n! automorphisms, so it takes a while to give you all of them. However, if you have posets that are more spread out by height and have fewer nodes at each height, it is significantly faster because it checks fewer automorphisms. This also has an in built poset checker when you enter a relations list. I am also working on it printing to file.